+++ wagtail_airtable/utils.py
@@ -4,8 +4,9 @@ Utility functions for wagtail-airtable.
 from django.conf import settings
 from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ObjectDoesNotExist
+from django.core.management.base import CommandError
 
-# TODO write tests for these
+
 def get_model_for_path(model_path):
     """
     Given an 'app_name.model_name' string, return the model class
@@ -15,21 +16,21 @@ def get_model_for_path(model_path):
         app_label, model_name
     ).model_class()
 
-def get_all_models() -> set:
+def get_all_models() -> list:
     """
     Get's all models from settings.AIRTABLE_IMPORT_SETTINGS.
 
-    Returns a set (unique and unordered list of models)
+    Returns a list of models.
     """
     airtable_settings = getattr(settings, "AIRTABLE_IMPORT_SETTINGS", {})
-    validated_models = set()
+    validated_models = []
     for label, model_settings in airtable_settings.items():
         if model_settings.get("AIRTABLE_IMPORT_ALLOWED", True):
             label = label.lower()
             if "." in label:
                 try:
                     model = get_model_for_path(label)
-                    validated_models.add(model)
+                    validated_models.append(model)
                 except ObjectDoesNotExist:
                     raise CommandError(
                         "%r is not recognised as a model name." % label
